requirements.txt
numpy<2.0.0
matplotlib
tqdm
pyyaml
einops
opencv-python
pylasadataset
wandb